+++ sdk/go/common/workspace/plugins.go
@@ -26,6 +26,7 @@ import (
 	"regexp"
 	"runtime"
 	"sort"
+	"strings"
 	"time"
 
 	"github.com/blang/semver"
@@ -190,8 +191,9 @@ func (info PluginInfo) Download() (io.ReadCloser, int64, error) {
 	// is hosted by Pulumi.
 	serverURL := info.ServerURL
 	if serverURL == "" {
-		serverURL = "https://api.pulumi.com/releases/plugins"
+		serverURL = "https://get.pulumi.com/releases/plugins"
 	}
+	serverURL = strings.TrimSuffix(serverURL, "/")
 
 	endpoint := fmt.Sprintf("%s/pulumi-%s-%s-v%s-%s-%s.tar.gz", serverURL, info.Kind, info.Name, info.Version, os, arch)
 	req, err := http.NewRequest("GET", endpoint, nil)